backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.4
selenium>=4.20
beautifulsoup4>=4.12
requests>=2.32
sentence-transformers>=3.0
chromadb>=0.5
python-dotenv>=1.0
PyMySQL>=1.1
openai>=1.37
